requirements.txt
numpy==1.24.1
scipy==1.9.3
tensorflow-cpu==2.9.3; sys_platform != 'darwin'
tensorflow-macos==2.9.2; sys_platform == 'darwin'
tensorflow-metal==0.5.0; sys_platform == 'darwin'
